Renamed public to guest .decorator.ts
