+++ src/app.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, Redirect } from '@nestjs/common';
 import { ApiHideProperty, ApiMovedPermanentlyResponse } from '@nestjs/swagger';
 import { AppService } from './app.service';
-import { Guest } from './auth/public.decorator';
+import { Guest } from './auth/guest.decorator';
 
 @Controller()
 export class AppController {
